+++ CHANGELOG.md
@@ -11,11 +11,30 @@
   [Marcelo Fabri](https://github.com/marcelofabri)
   [#1387](https://github.com/realm/SwiftLint/issues/1453)
 
+* Add `modificationDate(forFileAtPath:)` function requirement to
+  `LintableFileManager` protocol.  
+  [Victor Pimentel](https://github.com/victorpimentel)
+
+* Replace `configurationHash` parameters in all `LinterCache` initializers
+  with `configurationDescription` parameters.  
+  [Victor Pimentel](https://github.com/victorpimentel)
+
+* Remove `fileHash` parameter from `LinterCache.cache(violations:forFile:)`
+  and `LinterCache.violations(forFile:)` functions.  
+  [Victor Pimentel](https://github.com/victorpimentel)
+
+* Remove `hash` attribute from `Configuration` struct.  
+  [Victor Pimentel](https://github.com/victorpimentel)
+
 * Rename `ConditionalReturnsOnNewline` struct to
   `ConditionalReturnsOnNewlineRule` to match rule naming conventions.  
   [JP Simard](https://github.com/jpsim)
 
 ##### Enhancements
+
+* Cache linter results for files unmodified since the previous linter run.  
+  [Victor Pimentel](https://github.com/victorpimentel)
+  [#1184](https://github.com/realm/SwiftLint/issues/1184)
 
 * Add opt-in configurations to `generic_type_name`, `identifier_name` and
   `type_name` rules to allow excluding non-alphanumeric characters and names